+++ public/service-worker.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const APP_CACHE = 'gchat-pwa-v10';
+const APP_CACHE = 'gchat-pwa-v11';
 const APP_SHELL_ASSETS = [
   '/',
   '/index.html',